+++ app/src/Routes.tsx
@@ -16,7 +16,7 @@ const Routes = () => {
   const oidcConfig = useAppSelector(selectOidcConfig);
   const isBetaRelease = useAppSelector(selectIsBetaRelease);
   const routes = useRoutes(routeArray(oidcConfig.redirectPath, oidcConfig.popupRedirectPath));
-  const isAuthError = useAppSelector<Error | undefined>(selectAuthError);
+  const isAuthError = useAppSelector(selectAuthError);
 
   const renderRoutes = useMemo(() => {
     return isAuthError ? <ErrorLoginPage error={isAuthError} /> : routes;