+++ app/src/config/routes.tsx
@@ -2,13 +2,12 @@
 //
 // SPDX-License-Identifier: EUPL-1.2
 import { Box, Typography } from '@mui/material';
-import { AuthCallback } from '@opentalk/react-redux-appauth';
+import { AuthCallback, selectAuthError, selectIsAuthed, selectIsLoading, useAuth } from '@opentalk/react-redux-appauth';
 import i18next from 'i18next';
 import React, { ReactNode, useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
 import { To, RouteObject, useNavigate, Outlet, useParams } from 'react-router-dom';
 
-import LoginPopup from '../components/LoginPopup';
 import { useAppSelector } from '../hooks';
 import {
   SettingsProfilePage,
@@ -22,10 +21,11 @@ import {
   EventDetailsPage,
 } from '../pages/Dashboard';
 import RoomPage from '../pages/RoomPage';
-import { selectIsAuthenticated } from '../store/slices/userSlice';
+import { selectIsGuest } from '../store/slices/userSlice';
 import DashboardSettingsTemplate from '../templates/DashboardSettingsTemplate';
 import DashboardTemplate from '../templates/DashboardTemplate';
 import LobbyTemplate from '../templates/LobbyTemplate';
+import { sessionStorageItems } from './storage';
 
 const InvitePage = React.lazy(() => import('../pages/InvitePage'));
 
@@ -69,15 +69,36 @@ const Redirect = ({ to }: { to: To }) => {
   return null;
 };
 
+const WAIT_FOR_REDIRECT_BEFORE_SIGNIN = 500; //ms
+
 const ProtectedRoute = ({ children }: { children?: ReactNode }) => {
-  const isAuthenticated = useAppSelector(selectIsAuthenticated);
-  if (isAuthenticated) {
+  const { signIn } = useAuth();
+  const isAuthenticated = useAppSelector(selectIsAuthed);
+  const isAuthLoading = useAppSelector(selectIsLoading);
+  const isAuthError = useAppSelector(selectAuthError);
+  const isGuest = useAppSelector(selectIsGuest);
+  const startLogin = !isAuthenticated && !isAuthLoading && !isAuthError && !isGuest;
+
+  useEffect(() => {
+    if (startLogin) {
+      const timeout = setTimeout(async () => {
+        const redirectUrl = sessionStorage.getItem(sessionStorageItems.redirectUri);
+        if (!redirectUrl || redirectUrl === '/auth/popup_callback') {
+          sessionStorage.setItem(sessionStorageItems.redirectUri, window.location.pathname);
+        }
+        await signIn();
+      }, WAIT_FOR_REDIRECT_BEFORE_SIGNIN);
+      return () => clearTimeout(timeout);
+    }
+  }, [startLogin]);
+
+  if (isAuthenticated || isGuest) {
     if (children !== undefined) {
       return <>{children}</>;
     }
     return <Outlet />;
   }
-  return <LoginPopup />;
+  return null;
 };
 
 type CreateRoutes = (redirectUri: string, popUpRedirect: string) => RouteObject[];