+++ app/src/hooks/index.ts
@@ -1,9 +1,9 @@
 // SPDX-FileCopyrightText: OpenTalk GmbH <mail@opentalk.eu>
 //
 // SPDX-License-Identifier: EUPL-1.2
-import { TypedUseSelectorHook, useDispatch, useSelector } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 
-import { AppDispatch, RootState } from '../store';
+import { AppDispatch } from '../store';
 
 export { useDisplayName } from './ui';
 export * from './time';
@@ -12,4 +12,4 @@ export { default as useRemainingDurationOfTimer } from './useRemainingDurationOf
 
 // Use throughout your app instead of plain `useDispatch` and `useSelector`
 export const useAppDispatch = () => useDispatch<AppDispatch>();
-export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
+export const useAppSelector = useSelector;